+++ ruoyi-system-fx67ll/src/main/java/com/ruoyi/fx67ll/mahjong/service/impl/Fx67llMahjongRoomServiceImpl.java
@@ -6,6 +6,9 @@ import com.ruoyi.common.utils.DateUtils;
 import com.ruoyi.common.utils.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.ruoyi.common.core.domain.entity.SysUser;
+import com.ruoyi.system.service.ISysUserService;
+import com.ruoyi.common.exception.ServiceException;
 import com.ruoyi.fx67ll.mahjong.mapper.Fx67llMahjongRoomMapper;
 import com.ruoyi.fx67ll.mahjong.domain.Fx67llMahjongRoom;
 import com.ruoyi.fx67ll.mahjong.service.IFx67llMahjongRoomService;
@@ -21,6 +24,32 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
     @Autowired
     private Fx67llMahjongRoomMapper fx67llMahjongRoomMapper;
 
+    @Autowired
+    private ISysUserService userService;
+
+    // 管理员用户名
+    private static final String ADMIN_USERNAME = "fx67ll";
+
+    /**
+     * 权限校验：非管理员只能操作自己创建的数据
+     *
+     * @param room 麻将室对象
+     */
+    private void checkPermission(Fx67llMahjongRoom room) {
+        String currentUsername = SecurityUtils.getUsername();
+        // 管理员直接通过校验
+        if (ADMIN_USERNAME.equals(currentUsername)) {
+            return;
+        }
+        // 非管理员需校验创建者是否为当前用户
+        if (room == null) {
+            throw new ServiceException("数据不存在，无法操作！");
+        }
+        if (!currentUsername.equals(room.getCreateBy())) {
+            throw new ServiceException("您没有权限操作该数据，请立刻停止违规操作！");
+        }
+    }
+
     /**
      * 查询麻将室
      *
@@ -29,7 +58,10 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
      */
     @Override
     public Fx67llMahjongRoom selectFx67llMahjongRoomByMahjongRoomId(Long mahjongRoomId) {
-        return fx67llMahjongRoomMapper.selectFx67llMahjongRoomByMahjongRoomId(mahjongRoomId);
+        Fx67llMahjongRoom room = fx67llMahjongRoomMapper.selectFx67llMahjongRoomByMahjongRoomId(mahjongRoomId);
+        // 校验权限
+        checkPermission(room);
+        return room;
     }
 
     /**
@@ -40,6 +72,11 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
      */
     @Override
     public List<Fx67llMahjongRoom> selectFx67llMahjongRoomList(Fx67llMahjongRoom fx67llMahjongRoom) {
+        String currentUsername = SecurityUtils.getUsername();
+        // 非管理员只能查询自己创建的数据
+        if (!ADMIN_USERNAME.equals(currentUsername)) {
+            fx67llMahjongRoom.setCreateBy(currentUsername);
+        }
         return fx67llMahjongRoomMapper.selectFx67llMahjongRoomList(fx67llMahjongRoom);
     }
 
@@ -51,7 +88,12 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
      */
     @Override
     public List<Fx67llMahjongRoom> selectFx67llMahjongRoomListByUserId(Fx67llMahjongRoom fx67llMahjongRoom) {
+        String currentUsername = SecurityUtils.getUsername();
         fx67llMahjongRoom.setUserId(SecurityUtils.getUserId());
+        // 非管理员只能查询自己创建的数据
+        if (!ADMIN_USERNAME.equals(currentUsername)) {
+            fx67llMahjongRoom.setCreateBy(currentUsername);
+        }
         return fx67llMahjongRoomMapper.selectFx67llMahjongRoomList(fx67llMahjongRoom);
     }
 
@@ -63,7 +105,12 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
      */
     @Override
     public int insertFx67llMahjongRoom(Fx67llMahjongRoom fx67llMahjongRoom) {
-        fx67llMahjongRoom.setUserId(SecurityUtils.getUserId());
+        SysUser sysUser = userService.selectUserById(fx67llMahjongRoom.getUserId());
+        if (sysUser == null) {
+            throw new ServiceException("用户不存在，用户ID：" + fx67llMahjongRoom.getUserId());
+        }
+        // 新增数据
+        fx67llMahjongRoom.setUserName(sysUser.getUserName());
         fx67llMahjongRoom.setCreateBy(SecurityUtils.getUsername());
         fx67llMahjongRoom.setCreateTime(DateUtils.getNowDate());
         fx67llMahjongRoom.setUpdateBy(SecurityUtils.getUsername());
@@ -79,6 +126,12 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
      */
     @Override
     public int updateFx67llMahjongRoom(Fx67llMahjongRoom fx67llMahjongRoom) {
+        // 先查询原数据
+        Fx67llMahjongRoom oldRoom = fx67llMahjongRoomMapper.selectFx67llMahjongRoomByMahjongRoomId(
+                fx67llMahjongRoom.getMahjongRoomId());
+        // 校验权限
+        checkPermission(oldRoom);
+        // 更新信息
         fx67llMahjongRoom.setUpdateBy(SecurityUtils.getUsername());
         fx67llMahjongRoom.setUpdateTime(DateUtils.getNowDate());
         return fx67llMahjongRoomMapper.updateFx67llMahjongRoom(fx67llMahjongRoom);
@@ -92,6 +145,15 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
      */
     @Override
     public int deleteFx67llMahjongRoomByMahjongRoomIds(Long[] mahjongRoomIds) {
+        String currentUsername = SecurityUtils.getUsername();
+        // 非管理员需要逐个校验权限
+        if (!ADMIN_USERNAME.equals(currentUsername)) {
+            for (Long roomId : mahjongRoomIds) {
+                Fx67llMahjongRoom room = fx67llMahjongRoomMapper.selectFx67llMahjongRoomByMahjongRoomId(roomId);
+                // 校验权限
+                checkPermission(room);
+            }
+        }
         return fx67llMahjongRoomMapper.deleteFx67llMahjongRoomByMahjongRoomIds(mahjongRoomIds);
     }
 
@@ -103,6 +165,9 @@ public class Fx67llMahjongRoomServiceImpl implements IFx67llMahjongRoomService {
      */
     @Override
     public int deleteFx67llMahjongRoomByMahjongRoomId(Long mahjongRoomId) {
+        Fx67llMahjongRoom room = fx67llMahjongRoomMapper.selectFx67llMahjongRoomByMahjongRoomId(mahjongRoomId);
+        // 校验权限
+        checkPermission(room);
         return fx67llMahjongRoomMapper.deleteFx67llMahjongRoomByMahjongRoomId(mahjongRoomId);
     }
 }
